Fix #459
