+++ workflow/rules/build_electricity.smk
@@ -34,8 +34,6 @@ rule build_shapes:
 rule build_base_network:
     params:
         build_offshore_network=config["offshore_network"],
-        snapshots=config["snapshots"],
-        planning_horizons=config["scenario"]["planning_horizons"],
         model_topology=config["model_topology"]["include"],
     input:
         buses=DATA + "breakthrough_network/base_grid/bus.csv",
@@ -337,6 +335,8 @@ rule build_electrical_demand:
         demand_params=config["electricity"]["demand"],
         eia_api=config["api"]["eia"],
         profile_year=pd.to_datetime(config["snapshots"]["start"]).year,
+        planning_horizons=config["scenario"]["planning_horizons"],
+        snapshots=config["snapshots"],
     input:
         network=RESOURCES + "{interconnect}/elec_base_network.nc",
         demand_files=demand_raw_data,
@@ -361,6 +361,7 @@ rule build_sector_demand:
         planning_horizons=config["scenario"]["planning_horizons"],
         profile_year=pd.to_datetime(config["snapshots"]["start"]).year,
         eia_api=config["api"]["eia"],
+        snapshots=config["snapshots"],
     input:
         network=RESOURCES + "{interconnect}/elec_base_network.nc",
         demand_files=demand_raw_data,
@@ -505,6 +506,7 @@ rule add_demand:
     params:
         sectors=config["scenario"]["sector"],
         planning_horizons=config["scenario"]["planning_horizons"],
+        snapshots=config["snapshots"],
     input:
         network=RESOURCES + "{interconnect}/elec_base_network.nc",
         demand=demand_to_add,
