+++ workflow/scripts/_helpers.py
@@ -428,43 +428,6 @@ def mock_snakemake(rulename, **wildcards):
     return snakemake
 
 
-def test_column_datatypes_consistency(df):
-    """
-    Test if each column in a DataFrame has consistent datatypes using pandas
-    built-in methods.
-
-    Parameters:
-    df (pd.DataFrame): The DataFrame to test.
-
-    Returns:
-    list: A list of column names that do not have internally consistent datatypes.
-    """
-    inconsistent_columns = []
-    for column in df.columns:
-        # Use pandas infer_dtype to check for mixed types in the column
-        dtype = pd.api.types.infer_dtype(df[column], skipna=True)
-        if dtype.startswith("mixed"):
-            inconsistent_columns.append(column)
-    return inconsistent_columns
-
-
-def test_network_datatype_consistency(n):
-    """
-    Test if each component in a Network has consistent datatypes.
-    """
-    inconsistent_columns = {}
-    for component in n.components:
-        if component == "Network":
-            continue
-        col = test_column_datatypes_consistency(n.df(component))
-        if len(col) > 0:
-            inconsistent_columns[component] = col
-    if len(inconsistent_columns) > 0:
-        return f"Network has inconsistent datatypes in the following components: {inconsistent_columns}"
-    else:
-        return None
-
-
 def local_to_utc(group):
     import pytz
     from constants import STATE_2_TIMEZONE
@@ -890,3 +853,15 @@ def weighted_avg(df, values, weights):
     if valid.sum() == 0:
         return np.nan  # Return NaN if no valid entries
     return np.average(df[values][valid], weights=df[weights][valid])
+
+
+def get_multiindex_snapshots(
+    sns_config: dict[str, str],
+    invest_periods: list[int],
+) -> pd.MultiIndex:
+    sns = pd.DatetimeIndex([])
+    for year in invest_periods:
+        sns = sns.append(
+            get_snapshots(sns_config).map(lambda x: x.replace(year=year)),
+        )
+    return pd.MultiIndex.from_arrays([sns.year, sns])
